+++ src/choices-guide/component/question-group.jsx
@@ -36,7 +36,11 @@ export default class OpenStadComponentQuestionGroup extends OpenStadComponent {
       if ( !self.questionElements[i].isValid() ) {
         if (!scrollDone) {
           setTimeout( () => {
-            document.querySelector(`#${self.questionElements[i].config.divId}-content`).scrollIntoView({behavior: 'smooth'})
+            console.log(self.questionElements[i]);
+            console.log(self.questionElements[i].questionId);
+            let elem = document.querySelector(`#osc-question-${self.questionElements[i].questionId}-content`);
+            console.log(elem);
+            elem && elem.scrollIntoView({behavior: 'smooth'})
           }, 250 );
         }
         scrollDone = true;