+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Update ideas-overview, incl. add progress
 * Add warning when no choices guide was found
 * Simplify polyfills
+* Bugfix: scroll to unanswerd question
 
 ## v0.1.25
 * Add configurable placeholder image